requirements.txt
python-telegram-bot==20.6
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0